HC-242: Resolved file name too long issue (#85)
* HC-242: Resolve file name too long errors

This issue can be caused if you keep retrying a triaged dataset.
Therefore, this PR resolves that by basically checking to see if the
dataset being triaged is a triaged dataset. If so, then parse out the
job_id and use that when formulating the triaged dataset name

* HC-242: Resolve file name too long errors

This issue can be caused if you keep retrying a triaged dataset.
Therefore, this PR resolves that by basically checking to see if the
dataset being triaged is a triaged dataset. If so, then parse out the
job_id and use that when formulating the triaged dataset name

* added more logging; fallback to default regex if custom fails

* use underscore

* update unit teest

* update another unit test

* updated other unit tests

* fix unit test

* update unit tests

* remove custom triage regex

* add unit test to test triaging of a triaged dataset

* fix unit test

* remove main

Co-authored-by: Mike Cayanan <michael.d.cayanan@jpl.nasa.gov>