HC-300: improvements to harikiri and spot_termination_detector
Support the passing in of a configuration file to read in arguments to
the tool. Anything set on the command line will still take precedence
over what gets specified in a configuration file.

In regards to the spot_termination_detector script, an update was made
to immediately log a "marked_for_termination" event before it performs
the graceful shutdown.

Housekeeping Update:
- Fixed numerous flake8 errors